// ValuationClientApp/src/app/views/pages/models/userAreaOfSpecialization.model.ts
export interface UserAreaOfSpecialization {
    userAreaOfSpecializationId: number;
    userId: number;
    areaOfSpecializationName: string;
    isActive: boolean;
    createdDate: string;  // Can be Date type if needed
    createdById: number;
    modifiedDate: string;  // Can be Date type if needed
    modifiedById: number;
  }
  